+++ examples/lib/stories/effects/move_effect.dart
@@ -4,7 +4,6 @@ import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
 import 'package:flame/input.dart';
 
-import '../../commons/circle_component.dart';
 import '../../commons/square_component.dart';
 
 class MoveEffectGame extends FlameGame with TapDetector {
@@ -21,12 +20,10 @@ class MoveEffectGame extends FlameGame with TapDetector {
   @override
   Future<void> onLoad() async {
     await super.onLoad();
-    square = SquareComponent(size: 50)..position.setValues(200, 150);
+    square = SquareComponent(size: 50, position: Vector2(200, 150));
     add(square);
-    add(Component()
-      ..addAll([
-        for (final point in path) CircleComponent(radius: 3)..position = point
-      ]));
+    final pathMarkers = path.map((p) => CircleComponent(3, position: p));
+    addAll(pathMarkers);
   }
 
   @override